Feat : ex01

+++ cpp00/ex00/megaphone.cpp
@@ -6,7 +6,7 @@
  */
 void	changeToUppercaseFromLowercase (std::string &str)
 {
-	for (int i = 0 ; i < str.size() ; ++i)
+	for (size_t i = 0 ; i < str.size() ; ++i)
 		str[i] = static_cast<char>(std::toupper(str[i]));
 }
